requirements.txt
Django
djangorestframework
django-filter
drf-yasg
psycopg2-binary